Try to factor out parts of individual camera management
